+++ steam_web/Script/DTO/BuyOrderStatusAsset.cs
@@ -6,7 +6,7 @@ public class BuyOrderStatusAsset
 	[JsonPropertyName("accountid_seller")] public uint AccountidSeller { get; init; }
 	[JsonPropertyName("appid")] public uint AppId { get; init; }
 	[JsonPropertyName("assetid")] public ulong AssetId { get; init; }
-	[JsonPropertyName("contextid")] public byte ContextId { get; init; }
+	[JsonPropertyName("contextid")] public uint ContextId { get; init; }
 	[JsonPropertyName("currency")] public byte Currency { get; init; }
 	[JsonPropertyName("listingid")] public long ListingId { get; init; }
 	[JsonPropertyName("price_fee")] public uint PriceFee { get; init; }
